Add photo thumbs + popups
There is now a layercontrol in place with a checkbox
to view the photos (off by default). The only drawback
of including photos is that it takes 2 api calls per activity.
The api limit is 100 calls per 15 minutes. So this means
that my limit is around 48 activities per map, otherwise
I will exceed the api limit.