Only create the photos feature group if needed
In Summer 2022 the photos returned by Strava API
stopped having coordinates, so we would skip all
the photos when generating the map. Without
any pictures to show it's pointless to show
the "Show Photos" checkbox as it will not do
anything.

With this change, the feature group is only created
and added to the map once we find a photo with
coordinates.