Add scale to Thuderforest tiles template
The tiles support scaling:
https://www.thunderforest.com/tutorials/leaflet/

Hardcoding it to @2x would work as well, but with this,
Leaflet should inject the right value automatically.

This is what I'd been looking for for a while. In the past
I would I find the detect_retina option, but what that does
is different and the result is poor - it will just request 2x2
as many tiles and zoom out a bit to show all of them. The result
side effect is not just higher details, but also half the size
of the fonts which is not great.