Switch Stamen Terrain to new Stadia tile provider
The Stamen tiles are being moved to Stadia. Here's the announcement:
https://maps.stamen.com/stadia-partnership/

I created a free Stadia account and changed to the new tiles url
template. But somehow some of the tiles still always show a warning
that I should migrate (and that they will stop working end of 10/2023).
So let's see if it fixes itself or not.